AURION-Backend/requirements.txt

# requirements.txt

# Core Server
fastapi             # The web framework
uvicorn[standard]   # The server to run our app
sse-starlette       # For Server-Sent Events streaming
python-socketio     # For real-time WebSocket communication

# AI Clients
groq                # Groq API client
google-generativeai # Gemini API client
openai              # OpenAI Embedding API client
httpx               # Async HTTP client (good practice)
aiohttp             # Async HTTP client for web scraping
beautifulsoup4      # HTML parsing for web scraping

# Database Clients
pinecone            # Pinecone vector DB client
redis               # Redis client
neo4j               # Neo4j Graph database
motor               # MongoDB async driver
pymongo             # MongoDB driver
bcrypt              # Password hashing

# Tools & Utilities
pydantic-settings   # For loading our .env file
python-dotenv       # Manages .env variables
google-api-python-client # For YouTube/Search
google-auth         # Google OAuth authentication
google-auth-oauthlib # OAuth flow handling
google-auth-httplib2 # HTTP library for Google Auth
dateparser          # For parsing "tomorrow at 5pm"
apscheduler         # For scheduling reminders
fastapi-mail        # For sending emails
jinja2>=3.1,<3.2   # Templating engine required by fastapi-mail
aiosmtplib>=2.0    # SMTP client used by fastapi-mail
numpy               # For fallback embeddings
psutil              # For system monitoring (CPU, memory)
pyjwt               # For JWT token handling

# (keep all existing requirements)
beautifulsoup4
lxml